[*] Move wrapper code for fence and semaphore into subfolder
