+++ src/vulkan-renderer/wrapper/synchronization/fence.cpp
@@ -1,4 +1,4 @@
-#include "inexor/vulkan-renderer/wrapper/fence.hpp"
+#include "inexor/vulkan-renderer/wrapper/synchronization/fence.hpp"
 
 #include "inexor/vulkan-renderer/wrapper/device.hpp"
 #include "inexor/vulkan-renderer/wrapper/make_info.hpp"
@@ -7,7 +7,7 @@
 #include <stdexcept>
 #include <utility>
 
-namespace inexor::vulkan_renderer::wrapper {
+namespace inexor::vulkan_renderer::wrapper::synchronization {
 
 Fence::Fence(const Device &device, const std::string &name, const bool in_signaled_state)
     : m_device(device), m_name(name) {
@@ -42,4 +42,4 @@ VkResult Fence::status() const {
     return vkGetFenceStatus(m_device.device(), m_fence);
 }
 
-} // namespace inexor::vulkan_renderer::wrapper
+} // namespace inexor::vulkan_renderer::wrapper::synchronization
